Make animation rate a little more regular.
Resetting _clock after calling _pad.refresh(...)
gives a slightly better measure of the time that the frame has
been displayed. However, the time taken to update the terminal
may still cause the display to flicker with fast refresh rates.
For example, if, with a refresh rate of 0.1s, it takes 0.06s for
the terminal to refresh after _pad.refresh is called, then the
next call to _pad.refresh will occur 0.04 seconds after the actual
screen refresh, and the terminal 'could' occur much quicker this time,
resulting in a shorter than intended frame display time.